feat: Implement LivedataEvent and PulseEvent
LivedataEvent is a wrapper of MutableLivedata, that is aimed to be used in shared ViewModels to pass single event data between Fragments and Activities.
It does NOT guarantee that multiple observers won't get the same event and the event will be handled globally once.
PulseEvent is a subclass of LivedataEvent where the data in question is a boolean.
Both implementations are meant to reduce boiler plate code, which is the reason why this does not model a only-handled-once event.. yet